DAQ-104 Use display thread directly from ScriptingConnection.Java
The plotclient bundle is required by the GDA server to support plotting
from the Jython terminal, but commit ce03f352d29 introduced a dependency
on org.dawb.common.ui into the plotclient which, as a result, causes the
GDA server to depend on lots of client library code. Reverting line 298
in ScriptingConnection.Java to the original code allows this dependency
to be removed and hence the gda server to client dependency is minimised.

This does not result in any change of functionality as the original
change to the class just re-factored the operation to a common library
method.