I14-165 Better handling of no class attribute in PlotWindowManager
In some circumstances, such as when creating a view class with
parameters with <class class=...>, the "class" attribute of a view 
can be null. This would previously cause a null pointer exception
in PlotWindowManager.

This change:
(a) tries to find "class" in a child element if it is not at
the top level
(b) If it still cannot find the class, warns the user and ignores
the view rather than crashing.