DAWNSCI-6068: Update Python Path in PyDev interpreter info
On startup (unless -DDAWN_SKIP_UPDATE_PYTHON_PATH=true) the
PYTHONPATH for PyDev will be updated to ensure that the current 
scisoftpy (and Fabio) are included and older versions are excluded.

Includes factoring out common code from JythonCreator to be able to
reuse the same IStartup for both Jython and Python updates.

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>